requirements.txt
transformers>=4.43.0
datasets>=2.19.0
evaluate>=0.4.2
accelerate>=0.33.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
torch>=2.1.0
tensorboard>=2.14.0
pyyaml>=6.0.1
